+++ src/FileNotFoundException.php
@@ -11,6 +11,8 @@
 
 namespace Webmozart\Json;
 
+use RuntimeException;
+
 /**
  * Thrown when a file was not found.
  *
@@ -18,6 +20,6 @@ namespace Webmozart\Json;
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class FileNotFoundException extends \Exception
+class FileNotFoundException extends RuntimeException
 {
 }